+++ src/command/roll.rs
@@ -18,7 +18,7 @@ impl Command for Roll {
         "<dice>. Examples: [\x02roll\x02 d20 + 4 - 2d6!], [\x02roll\x02 3dF-2], [\x02roll\x02 2d6>3 - 1d4].".to_owned() 
     }
     fn fits(&self, size: usize) -> bool { size >= 1 }
-    fn auth(&self) -> i32 { 0 }
+    fn auth(&self) -> u8 { 0 }
 
     fn run(&mut self, args: &[&str], _: &Context, _: &mut Db) -> Outcome {
         let content = args.join(" ");
